requirements.txt
alembic==1.7.7
anyio==3.5.0
asgiref==3.5.0
beautifulsoup4==4.10.0
bs4==0.0.1
certifi==2021.10.8
charset-normalizer==2.0.12
click==8.0.4
et-xmlfile==1.1.0
fastapi==0.71.0
greenlet==1.1.2
h11==0.13.0
idna==3.3
Jinja2==3.0.3
lxml==4.8.0
Mako==1.2.0
MarkupSafe==2.1.1
numpy==1.22.3
openpyxl==3.0.9
pandas==1.4.1
pydantic==1.9.0
python-dateutil==2.8.2
pytz==2021.3
requests==2.27.1
six==1.16.0
sniffio==1.2.0
soupsieve==2.3.1
SQLAlchemy==1.4.32
starlette==0.17.1
typing_extensions==4.1.1
urllib3==1.26.9
uvicorn==0.17.6
